Added 7 day incidence reporting
The 7-day-incidence is the number of cases per 100,000 inhabitants. This is a very important number, used to determine the Coronavirus Ampel colour for example, as well as by companies I work with to determine if meetings can happen.

I wasn't able to find an easy way to pull this into my data, so I just calculated it myself.